+++ CHANGES.md
@@ -15,6 +15,7 @@ Changes in 6.0
 - Add Warehouse Header Javascripts WarehouseHeaderJS() function in Warehouse.php
 - HTML Do not close body & html (footer) if not AJAX in Warehouse.php
 - Support multiple values in SelectInput() in Inputs.php
+- Remove FirstLoginLoadJSCSS() & use form `target="_top"` instead in FirstLogin.fnc.php
 
 Changes in 5.9.3
 ----------------